+++ Material.Blazor/Components/Grid/MBGridTextColorSpecification.cs
@@ -6,7 +6,7 @@ namespace Material.Blazor
     {
         public Color BackgroundColor { get; set; }
         public Color ForegroundColor { get; set; }
-        public bool Supress { get; set; } = false;
+        public bool Suppress { get; set; } = false;
         public string Text { get; set; }
     }
 }